+++ packages/framework/src/Form/DisplayOnlyUrlType.php
@@ -1,0 +1,29 @@
+<?php
+
+namespace Shopsys\FrameworkBundle\Form;
+
+use Symfony\Component\Form\AbstractType;
+use Symfony\Component\OptionsResolver\OptionsResolver;
+
+class DisplayOnlyUrlType extends AbstractType
+{
+    /**
+     * @param \Symfony\Component\OptionsResolver\OptionsResolver $resolver
+     */
+    public function configureOptions(OptionsResolver $resolver)
+    {
+        $resolver
+            ->setDefaults([
+                'mapped' => false,
+                'required' => false,
+                'attr' => [
+                    'readonly' => 'readonly',
+                ],
+                'data' => [
+                    'route_name' => '',
+                    'id' => null,
+                    'name' => '',
+                ],
+            ]);
+    }
+}